moved wireless discovery to its own window
