+++ Menus/MainWindow.xaml
@@ -5,7 +5,7 @@
         xmlns:lvc="clr-namespace:LiveChartsCore.SkiaSharpView.WPF;assembly=LiveChartsCore.SkiaSharpView.WPF"
         Title="Networks"
         ShowInTaskbar="False"
-        Width="350" Height="400"
+        Width="350" Height="450"
         Background="Transparent"
         AllowsTransparency="True"
         WindowStyle="None"
@@ -115,31 +115,7 @@
                         </DataTemplate>
                         </ItemsControl.ItemTemplate>
                     </ItemsControl>
-                    <Separator Margin="0,5,0,5"/>
-                    <StackPanel Name="AvailableNetworks">
-                        <TextBlock Text="Wireless Networks" Foreground="White" FontSize="16" Margin="0,0,0,10"/>
-                        <StackPanel Name="AvailableNetworksPanel" Margin="10">
-                            <TextBlock Text="No Wi-Fi adapter detected." Foreground="LightGray" Name="NoWiFiTextBlock"/>
-                        </StackPanel>
-                        <ItemsControl Name="WiFiNetworkList">
-                            <ItemsControl.ItemTemplate>
-                                <DataTemplate>
-                                    <Expander Header="{Binding SSID}" Foreground="White" FontSize="14" Margin="5">
-                                        <StackPanel Margin="10,0,0,0">
-                                            <TextBlock Text="{Binding NetworkType}" Foreground="LightGray"/>
-                                            <TextBlock Text="{Binding Authentication}" Foreground="LightGray"/>
-                                            <TextBlock Text="{Binding Encryption}" Foreground="LightGray"/>
-                                            <TextBlock Text="{Binding SignalStrength}" Foreground="LightGray"/>
-                                            <Button Content="Connect" Click="ConnectToWiFi_Click"
-                                    Tag="{Binding SSID}" 
-                                    Background="Gray" Foreground="White" 
-                                    Width="80" Margin="0,5,0,0"/>
-                                        </StackPanel>
-                                    </Expander>
-                                </DataTemplate>
-                            </ItemsControl.ItemTemplate>
-                        </ItemsControl>
-                    </StackPanel>
+
                 </StackPanel>
             </ScrollViewer>
         </Grid>
