+++ Menus/WirelessWindow.xaml
@@ -26,8 +26,61 @@
             </Grid.RowDefinitions>
             <ScrollViewer Grid.Row="0" VerticalScrollBarVisibility="Auto">
                 <StackPanel>
-                    <TextBlock Text="Wi-Fi Networks" Foreground="White" FontSize="16" Margin="0,0,0,10"/>
-                    <StackPanel Name="NetworkListPanel" Margin="10"/>
+                    <Grid VerticalAlignment="Top" HorizontalAlignment="Left" ShowGridLines="True">
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="auto"></ColumnDefinition>
+                            <ColumnDefinition Width="auto"></ColumnDefinition>
+                            <ColumnDefinition Width="auto"></ColumnDefinition>
+                            <ColumnDefinition Width="auto"></ColumnDefinition>
+                        </Grid.ColumnDefinitions>
+                        <Grid.RowDefinitions>
+                            <RowDefinition Height="auto"></RowDefinition>
+                        </Grid.RowDefinitions>
+                        
+                        <Button Name="Home" Content="Home" Click="HomeButton_Click"
+                                Width="32" Height="32"
+                                Background="Transparent" Foreground="White"
+                                BorderThickness="0" Cursor="Hand"
+                                Grid.Column="0"
+                                Grid.Row="0"/>
+                        <Button Name="RefreshButton" Content="Refresh" Click="RefreshWiFi_Click"
+                                Width="32" Height="32"
+                                Background="Transparent" Foreground="White"
+                                BorderThickness="0" Cursor="Hand"
+                                Grid.Column="1"
+                                Grid.Row="0"/>
+                        
+                        <TextBlock Text="Wi-Fi Networks" Foreground="White" 
+                                   FontSize="16"
+                                   Grid.Column="2"
+                                   Grid.Row="0"/>
+                    </Grid>
+                    <StackPanel Name="AvailableNetworks">
+                        <StackPanel Name="AvailableNetworksPanel" Margin="10">
+                            <TextBlock Text="No Wi-Fi adapter detected." Foreground="LightGray" Name="NoWiFiTextBlock"/>
+                        </StackPanel>
+                        <ItemsControl Name="WiFiNetworkList">
+                            <ItemsControl.ItemTemplate>
+                                <DataTemplate>
+                                    <Expander Header="{Binding SSID}" Foreground="White" FontSize="14" Margin="5">
+                                        <Expander.Resources>
+                                            <Style TargetType="Expander" BasedOn="{StaticResource ExpanderStyle1}"/>
+                                        </Expander.Resources>
+                                        <StackPanel Margin="10,0,0,0">
+                                            <TextBlock Text="{Binding NetworkType}" Foreground="LightGray"/>
+                                            <TextBlock Text="{Binding Authentication}" Foreground="LightGray"/>
+                                            <TextBlock Text="{Binding Encryption}" Foreground="LightGray"/>
+                                            <TextBlock Text="{Binding SignalStrength}" Foreground="LightGray"/>
+                                            <Button Content="Connect" Click="ConnectToWiFi_Click"
+                                    Tag="{Binding SSID}" 
+                                    Background="Gray" Foreground="White" 
+                                    Width="80" Margin="0,5,0,0"/>
+                                        </StackPanel>
+                                    </Expander>
+                                </DataTemplate>
+                            </ItemsControl.ItemTemplate>
+                        </ItemsControl>
+                    </StackPanel>
                 </StackPanel>
             </ScrollViewer>
         </Grid>
